+++ packages/flame/src/Input/Input.tsx
@@ -321,15 +321,6 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(
     const nextStatus = status && typeof status === 'object' ? status.type : (status as StatusType);
     const nextMessage = status && typeof status === 'object' ? status.message : statusMessage;
 
-    if (typeof status === 'object') {
-      // eslint-disable-next-line no-console
-      console.warn(`
-Using status as an object will be deprecated in the next version of Flame.
-Please prefer passing the status.type to the status prop directly and
-pass the status.message to the statusMessage prop.
-`);
-    }
-
     return (
       <React.Fragment>
         {label && (